requirements.txt
cycler==0.10.0
jedi==0.11.0
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
pandas==0.21.0
parso==0.1.0
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
torch==0.2.0.post3
torchvision==0.1.9
